refactor init function to simplify date range initialization and add Byte function for random byte generation

+++ init.go
@@ -16,8 +16,8 @@ var dtMin, dtMax int64
 
 func init() {
 	rnd = rand.New(rand.NewSource(time.Now().UnixNano()))
-	dtMin = int64(time.Date(1, 1, 1, 0, 0, 0, 0, time.UTC).Unix())
-	dtMax = int64(time.Date(9999, 12, 31, 23, 59, 59, 0, time.UTC).Unix())
+	dtMin = time.Date(1, 1, 1, 0, 0, 0, 0, time.UTC).Unix()
+	dtMax = time.Date(9999, 12, 31, 23, 59, 59, 0, time.UTC).Unix()
 }
 
 func Reseed() {
